Move prompt service into own folder
