+++ apps/desktop/src/lib/prompt/promptService.ts
