+++ apps/desktop/src/routes/+layout.ts
@@ -3,13 +3,13 @@ import { AIService } from '$lib/ai/service';
 import { initAnalyticsIfEnabled } from '$lib/analytics/analytics';
 import { PostHogWrapper } from '$lib/analytics/posthog';
 import { CommandService } from '$lib/backend/ipc';
-import { PromptService } from '$lib/backend/prompt';
 import { Tauri } from '$lib/backend/tauri';
 import { loadAppSettings } from '$lib/config/appSettings';
 import { GitConfigService } from '$lib/config/gitConfigService';
 import { FileService } from '$lib/files/fileService';
 import { HooksService } from '$lib/hooks/hooksService';
 import { ProjectsService } from '$lib/project/projectsService';
+import { PromptService } from '$lib/prompt/promptService';
 import { RemotesService } from '$lib/remotes/remotesService';
 import { RustSecretService } from '$lib/secrets/secretsService';
 import { TokenMemoryService } from '$lib/stores/tokenMemoryService';
